requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
PyGithub==2.1.1
python-dotenv==1.0.0
anthropic>=0.40.0
aiohttp==3.9.1
pytz>=2024.1
